Adds scoring systems in GameManager.cs, ball speed increased for each volley
Adds scoring system in GameManager.cs. A point is awarded to player one
or two depending on if the ball passes a paddle and hits the boundary on
the x-axis. A player UI is instantiated on Start and the player UI is
updated with scores for both players and keeps track of the volley
count. The ball speed is increased for each consecutive volley.